requirements.txt
django>=6.0.3
djangorestframework>=3.15.2
django-environ>=0.11.2
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=24.2
drf-spectacular>=0.27.2
django-polymorphic>=3.1.0
django-allauth>=65.4.1
dj-rest-auth>=6.0.0
requests>=2.32.3
faker>=33.3.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
redis>=5.2.1
celery>=5.4.0
pytesseract>=0.3.13
Pillow>=11.0.0
